flask-api/requirements.txt
flask
flask-cors
mysql-connector-python
pillow
matplotlib
numpy
tensorflow